fix(shorebird_cli): include link % and debug info in patch summary
