+++ packages/shorebird_cli/lib/src/commands/patch/patch_command.dart
@@ -89,8 +89,7 @@ of the iOS app that is using this module.''',
         'debug-linker',
         defaultsTo: true,
         help: 'Collects linker diagnostic information to help troubleshoot low '
-            'link percentages. File is saved to build/$linkDebugInfoFileName. '
-            'iOS only.',
+            'link percentages (iOS only.)',
       )
       ..addFlag(
         'dry-run',
@@ -298,6 +297,12 @@ NOTE: this is ${styleBold.wrap('not')} recommended. Asset changes cannot be incl
         '🟠 Track: ${lightCyan.wrap('Staging')}'
       else
         '🟢 Track: ${lightCyan.wrap('Production')}',
+      if (patcher.linkPercentage != null)
+        '''🔗 Running ${lightCyan.wrap('${patcher.linkPercentage!.toStringAsFixed(1)}%')} on CPU''',
+      if (results['debug-linker'] == true &&
+          (patcher.linkPercentage != null &&
+              patcher.linkPercentage! < Patcher.minLinkPercentage))
+        '''🔍 Debug Info: ${lightCyan.wrap(patcher.debugInfoFile.path)}''',
     ];
 
     logger.info(
